feat: run linux release on an old linux container
Building portable binaries on linux is a bit of a dumpster fire. When
you build on a modern linux version then you dynamically link against
a modern version of GLIBC and the binary is not compatible with older
linux releases.

To solve this you must build against an older version of GLIBC. This
then uses the version of GLIBC on the target as modern GLIBCs are
backward compatible.

There are a couple of different ways this can be achieved:

1. run the release build on an older version of linux (this approach).
   This approach was very common before good cross-compilers were
   available. This makes it compatible with GLIBC 2.26 (August 2017).
2. use a cross compiler. This is what cargo-binstall does, it uses
   cargo-zigbuild to build a GLIBC 2.17 compatible binary (December
   2012).
3. statically link GLIBC or MUSL LIBC. This is quite hard to get right
   and can have unintended consequences. In googling this I see a lot of
   people suggest "just statically link MUSL" and a lot of replies
   saying this should be avoided for a number of good reasons.

This builds with Amazon Linux 2 which gives GLIBC 2.26 or later
compatibility. This is the runtime for AWS Lambda so it makes a lot of
sense to have compatibility for this. The gold-standard approach for
rust is to use cargo-zigbuild to use zig as a cross-compiling linker.
This is what cargo-binstall and cargo-lambda do. It's more complex to
setup so I was reluctant to take this approach without agreement that
it was the right way forward. Happy to give it a go if preferred.